move event consumer and footballData to lib
